requirements.txt
django 
djangorestframework
python-decouple
factory-boy
djangorestframework-simplejwt
django-filter